+++ scripts/run.sh
@@ -53,6 +53,8 @@ shoot light screenshot-capture-light -pulse-capture
 shoot dark  screenshot-capture-dark  -pulse-capture
 shoot light screenshot-pv-light -pulse-capture-pv
 shoot dark  screenshot-pv-dark  -pulse-capture-pv
+shoot light screenshot-zurueck-light -pulse-capture-step2
+shoot dark  screenshot-zurueck-dark  -pulse-capture-step2
 shoot light screenshot-verlauf-light -pulse-verlauf
 shoot dark  screenshot-verlauf-dark  -pulse-verlauf
 shoot light screenshot-leer-light -pulse-empty